requirements.txt
imblearn==0.0
joblib==1.4.2
matplotlib==3.7.1
numpy==1.24.3
plotly==6.0.0
pandas==2.0.2
Pillow==9.5.0
scikit_learn==1.3.0
seaborn==0.13.2
streamlit==1.43.1
